# swagger/yaml/swagger-workflow.yaml
swagger: "2.0"
info: 
   title: "API for VMS Workflow Managment"
   description: "VMS Workflow Management APIs"
   version: 1.0.0
basePath: /api/v1
paths:
...
  /processInstance:
    post:
      tags:
      - "process instance managment"
      summary: "创建流程实例"
      description: ""
      operationId: createProcess
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: startProcessInstanceReq
        in: body
        required: true
        schema:
          $ref: "#/definitions/StartProcessInstanceReqVO"
      responses:
        200:
          description: "create instance successfully"
          schema:
            $ref: "#/definitions/StandardResponse"
  /processInstance/instanceKey/{key}:
    get:
      tags:
      - "process instance managment"
      summary: "根据instance key获取列表"
      description: ""
      operationId: queryProcessInstanceByKey
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: key
        in: path
        required: true
        type: string
      responses: 
        200:
          description: "get process instance list"
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
              data:
                type: array
                items:
                  $ref: "#/definitions/ProcessInstanceEntity"
  /processInstance/{id}:
    get:
      tags:
      - "process instance managment"
      summary: "根据instance id查看流程实例"
      description: ""
      operationId: queryProcessInstanceByID
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses: 
        200:
          description: "get process instance list"
          schema:
            $ref: "#/definitions/ProcessInstanceEntity"
    delete:
      tags:
      - "process instance managment"
      summary: "根据instance id删除流程实例"
      description: ""
      operationId: deleteProcessInstanceByID
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: "delete process instance successfully"
          schema:
            $ref: "#/definitions/StandardResponse"       
  /processInstance/historic/{id}:
    get:
      tags:
      - "process instance managment"
      summary: "根据ID查询 process instance的历史归档记录"
      description: ""
      operationId: queryTaskHistoricalData
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses: 
        200:
          description: "get instance historical task list"
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
              data:
                type: array
                items:
                  $ref: "#/definitions/HistroicTaskInstance"
...
definitions: 
...
  ProcessInstanceEntity:
    properties:
      processDefinitionId:
        type: integer
      processDefinitionName:
        type: string
      processDefinitionKey:
        type: string
      processDefinitionVersion:
        type: integer
      deploymentId:
        type: string
      businessKey:
        type: string
      isSuspended:
        type: boolean
      tenantId:
        type: string
      name:
        type: string
      startUserId:
        type: string
  HistroicTaskInstance:
    properties:
      deleteReason:
        type: string
      startTime:
        type: string
        format: date-time
      endTime:
        type: string
        format: date-time
      durationInMillis:
        type: integer
        format: int64
      workTimeInMillis:
        type: integer
        format: int64
      claimTime:
        type: string
        format: date-time
...
  StartProcessInstanceReqVO:
    properties:
      firstResult:
        type: integer
      maxResults:
        type: integer
      assignee:
        type: string
      candidateUser:
        type: string
      candidateGroup:
        type: string
...
  StandardResponse:
    required: 
       - code
       - message
       - data
    properties:
      code:
        type: integer
      message:
        type: string
      data:
        type: object
...